matching_api/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-core
langgraph
supabase
python-dotenv
pydantic